Fix critical bugs and modernize Docker build

+++ include/sendsys.h
@@ -1,6 +1,8 @@
 #ifndef _SENDSYS_H
 #define _SENDSYS_H
 
+#include <stdint.h>
+
 /* Special destinations for send message */
 #define DEST_ALL  (-32767)
 #define NOBODY    (-32766)    /* Value of x1 and x2 below if ignored */
@@ -55,16 +57,16 @@ void sillycom(char *txt);
 void sillytp(int per, char *msg);
 
 void send_g_msg(int destination,
-		char *func(int plx, int arg, char *t),
-		int  arg,
+		char *func(int plx, intptr_t arg, char *t),
+		intptr_t arg,
 		char *text);
 
 #ifdef VARGS
 void sendf(int destination,char *format,...);
 
 void gsendf(int destination,
-	    char *func(int plx, int arg, char *text),
-	    int arg,
+	    char *func(int plx, intptr_t arg, char *text),
+	    intptr_t arg,
 	    char *format,...);
 
 void send_msg(int destination,
