+++ src/sendsys.c
@@ -138,8 +138,8 @@ test_rcv (int player,		/* Who to send to */
  */
 void
 send_g_msg (int destination,	/* Where to send to */
-	    char *func (int plx, int arg, char *t),	/* Test function */
-	    int arg,		/* Argument to test */
+	    char *func (int plx, intptr_t arg, char *t),	/* Test function */
+	    intptr_t arg,	/* Argument to test */
 	    char *text)
 {				/* Text to send */
   char *t;
@@ -192,7 +192,7 @@ send_g_msg (int destination,	/* Where to send to */
 }
 
 char *
-check_send_msg (int plx, int a, char *t)
+check_send_msg (int plx, intptr_t a, char *t)
 {
   struct _send_msg_box *b = (struct _send_msg_box *) a;
 
@@ -227,7 +227,7 @@ send_msg (int destination,	/* Where to send to */
   va_start (pvar, format);
   vsprintf (bf, bb, pvar);
   va_end (pvar);
-  send_g_msg (destination, check_send_msg, (int) &b, bf);
+  send_g_msg (destination, check_send_msg, (intptr_t) &b, bf);
 }
 
 
@@ -245,8 +245,8 @@ sendf (int destination, char *format,...)
 
 void
 gsendf (int destination,
-	char *func (int plx, int arg, char *text),
-	int arg,
+	char *func (int plx, intptr_t arg, char *text),
+	intptr_t arg,
 	char *format,...)
 {
   char b[2048];
@@ -284,7 +284,7 @@ send_msg (int destination,	/* Where to send to */
   b.lang = NFL_ENGLISH;
   bb = format;
   sprintf (bf, bb, a1, a2, a3, a4, a5, a6, a7, a8, a9);
-  send_g_msg (destination, check_send_msg, (int) &b, bf);
+  send_g_msg (destination, check_send_msg, (intptr_t) &b, bf);
 }
 
 void
@@ -299,8 +299,8 @@ sendf (int destination, char *format, int a1, int a2, int a3, int a4, int a5,
 
 void
 gsendf (int destination,
-	char *func (int plx, int arg, char *text),
-	int arg,
+	char *func (int plx, intptr_t arg, char *text),
+	intptr_t arg,
 	char *format, int a1, int a2, int a3, int a4, int a5, int a6, int a7,
 	int a8, int a9)
 {
@@ -342,5 +342,5 @@ lsend_msg (int destination,	/* Where to send to */
   vsprintf (bf, bb, pvar);
   va_end (pvar);
 
-  send_g_msg (destination, check_send_msg, (int) &b, bf);
+  send_g_msg (destination, check_send_msg, (intptr_t) &b, bf);
 }
